requirements.txt
# Core LangChain + Langfuse (OBRIGATÓRIO - Tracking de custos)
langchain>=0.1.0
langchain-openai>=0.0.5
langfuse>=3.0.0,<4.0.0
python-dotenv>=1.0.0
ulid-py>=1.1.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
polars>=0.19.0

# Machine Learning - Core
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
catboost>=1.2.0
imbalanced-learn>=0.11.0

# Deep Learning (opcional - escolher um)
torch>=2.0.0
# tensorflow>=2.13.0

# Graph Analysis
networkx>=3.0

# Geospatial
geopandas>=0.14.0
shapely>=2.0.0

# NLP
spacy>=3.7.0
transformers>=4.30.0

# Multi-Agent Framework
langgraph>=0.0.20

# Utilities
tqdm>=4.65.0
joblib>=1.3.0
pyyaml>=6.0

# Explainability
shap>=0.42.0

# Monitoring
mlflow>=2.8.0

# Testing
pytest>=7.4.0
